refactor(reader): address review of IoBounds bounds typing (ADR 0003 Phase E)
Three follow-ups from PR review:

- ScanIterator.readFlatStats: degrade to ArrayStats.empty() on a malformed
  stats segment instead of throwing VortexException, matching
  VortexReader.readFlatStats. Stats are an optional zone-map pruning
  optimization — a corrupt stats segment must not abort the scan. Also adds
  the missing segIdx bounds guard the other reader already had.
- FlatSegmentDecoder: route the buffer-count allocation through
  IoBounds.checkCount before new MemorySegment[numBuffers].
- Date/Time/Timestamp/Uuid extension decoders: the decodeAll loops counted
  with a long index over an int bound (n); narrow to int i.

reader 705 green; checkstyle + javadoc clean.

+++ reader/src/main/java/io/github/dfa1/vortex/reader/extension/TimeExtensionDecoder.java
@@ -76,12 +76,12 @@ public final class TimeExtensionDecoder implements ExtensionDecoder {
         int n = IoBounds.toIntSize(storage.length());
         List<LocalTime> out = new ArrayList<>(n);
         if (storage instanceof MaskedArray masked) {
-            for (long i = 0; i < n; i++) {
+            for (int i = 0; i < n; i++) {
                 out.add(masked.isValid(i) ? decode(ext, masked.inner(), i) : null);
             }
             return out;
         }
-        for (long i = 0; i < n; i++) {
+        for (int i = 0; i < n; i++) {
             out.add(decode(ext, storage, i));
         }
         return out;
